Release 1.2.0: add jwks_refresh_interval option; extract Controller::ErrorHandling
Add the `jwks_refresh_interval` configuration option end-to-end so apps can
forward verikloak 1.1.0's JWKS revalidation throttle:
- Configuration accessor (default nil) and middleware_options forwarding,
  keeping the .compact nil-omission so the core default (60s) applies when
  unset and an explicit 0 (revalidate every request) is preserved
- Railtie::CONFIG_KEYS entry, README config-table row + VERIKLOAK_JWKS_
  REFRESH_INTERVAL ENV mapping, generator initializer template comment
- Specs: accessor/forwarding in configuration_spec, and the core contract
  example now sets the key so the real Middleware constructor is exercised

Extract the controller rescue handlers and 500-error logging into
Verikloak::Rails::Controller::ErrorHandling (behavior unchanged) so the
concern stays under the default Metrics/ModuleLength limit, and revert the
ModuleLength relaxation that had been added to .rubocop.yml.

Bump version to 1.2.0 and update CHANGELOG / Gemfile.lock.
