+++ lib/verikloak/rails/version.rb
@@ -2,6 +2,6 @@
 
 module Verikloak
   module Rails
-    VERSION = '1.1.0'
+    VERSION = '1.2.0'
   end
 end
